upload

+++ app/content/pegawai_hapus.php
@@ -1,22 +1,34 @@
 <?php
-if(!defined('INDEX')) die();
+if (!defined('INDEX')) die();
 
-$id   = $_GET['id'];
-$foto = $_GET['foto'];
+if (isset($_GET['id'])) {
+    $id = $_GET['id'];
 
-if(file_exists("images/$foto")){
-    unlink("images/$foto");
+    // Ambil data pegawai untuk mendapatkan nama file foto
+    $query = "SELECT foto FROM pegawai WHERE id_pegawai = '$id'";
+    $result = mysqli_query($con, $query);
+
+    if ($result && mysqli_num_rows($result) > 0) {
+        $data = mysqli_fetch_assoc($result);
+        $foto = $data['foto'];
+
+        // Pastikan $foto tidak kosong sebelum mencoba menghapusnya
+        if (!empty($foto) && file_exists("images/$foto")) {
+            unlink("images/$foto");
+        }
+
+        // Hapus data pegawai dari database
+        $delete = mysqli_query($con, "DELETE FROM pegawai WHERE id_pegawai = '$id'");
+
+        if ($delete) {
+            echo "<script>alert('Data pegawai berhasil dihapus!'); window.location='?hal=pegawai';</script>";
+        } else {
+            echo "<script>alert('Gagal menghapus data pegawai!'); window.location='?hal=pegawai';</script>";
+        }
+    } else {
+        echo "<script>alert('Data pegawai tidak ditemukan atau sudah dihapus sebelumnya!'); window.location='?hal=pegawai';</script>";
+    }
+} else {
+    echo "<script>alert('ID pegawai tidak valid!'); window.location='?hal=pegawai';</script>";
 }
-
-$query  = "DELETE FROM pegawai WHERE id_pegawai = '$id'";
-$result = mysqli_query($con,$query);
-
-if($result){
-    echo "Data pegawai berhasil dihapus!";
-    echo "<meta http-equiv='refresh' content='1; url=?hal=pegawai'>";
-}else{
-    echo "Tidak dapat menghapus data pegawai!<br>";
-    echo mysqli_error();
-}
-
 ?>
